+++ core/src/model/NonFoodGood.java
@@ -1,11 +1,13 @@
 package model;
 
 import util.DateUtil;
+import java.math.BigDecimal;
+import java.time.LocalDate;
 
 public class NonFoodGood extends Good {
-    private static final double MARKUP_PERCENTAGE = 0.30;
+    private static final BigDecimal MARKUP_PERCENTAGE = 0.30;
 
-    public NonFoodGood(String id, String name, double unitDeliveryPrice, LocalDate expirationDate, int quantity) {
+    public NonFoodGood(String id, String name, BigDecimal unitDeliveryPrice, LocalDate expirationDate, int quantity) {
         super(id, name, unitDeliveryPrice, "Non-Food", expirationDate, quantity);
     }
